feat: read ldr and ultrasonic

+++ arturitov2/ar2ito/Ultrasonic.cpp
@@ -9,7 +9,7 @@ Ultrasonic::Ultrasonic(int trigPin, int echoPin)
     pinMode(this->echoPin, INPUT);
 }
 
-int Ultrasonic::getDistance()
+float Ultrasonic::getDistance()
 {
     digitalWrite(this->trigPin, LOW);
     delayMicroseconds(2);
@@ -18,6 +18,6 @@ int Ultrasonic::getDistance()
     digitalWrite(this->trigPin, LOW);
 
     float duration = pulseIn(this->echoPin, HIGH);
-    float distance = duration * 0.0343 / 2;
+    float distance = duration * 0.0343 / 2.0;
     return distance;
 }
